refactor: move ChannelButton to its own file
