+++ src/components/common/atoms/index.tsx
@@ -1,4 +1,5 @@
 export {Button} from './Button';
+export {ChannelButton} from './ChannelButton';
 export {Checkbox} from './Checkbox';
 export {ContextButton} from './ContextButton';
 export {CopyIDButton} from './CopyIDButton';
