+++ src/components/navigation/ChannelList.tsx
@@ -6,9 +6,9 @@ import FastImage from 'react-native-fast-image';
 
 import {Channel, Server} from 'revolt.js';
 
-import {ChannelButton, app, client} from '../../Generic';
+import {app, client} from '../../Generic';
 import {styles} from '../../Theme';
-import {Text} from '../common/atoms';
+import {ChannelButton, Text} from '../common/atoms';
 const Image = FastImage;
 
 type ChannelListProps = {
